nest-ai-proto/Services/Backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML - Whisper for transcription
openai-whisper==20231117
torch==2.1.1
torchaudio==2.1.1

# Audio/Video processing
ffmpeg-python==0.2.0

# Task queue and caching
celery==5.3.4
redis==5.0.1

# Database (optional - using SQLite for simplicity)
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
